Make possible to skip first download
If the repo is corrupted or anything wrong happened, we should
have a way to override the download from the external source
first.

This makes possible to have in the CI a variable named
DOWNLOAD_FIRST. If this variable is set to False (or similar),
it will skip the download.