Remove lxc artifact config from python build
In the python artifact build we cannot use container
artifacts. They are created in the next step afterwards.
However the AIO bootstrap process puts configuration in
place to ensure that container artifacts are used when
doing production deployments.

This patch does a simple removal of those settings so
that the LXC default artifacts are used instead.

Connects https://github.com/rcbops/u-suk-dev/issues/1360