Adapt artifact job with new repo build
Repo build process has changed, and the urls are now different.
If we keep the same process in repo build and publish with the
new directory structure, we also need our checks to use that
new directory structure.

This should be enough to do the trick for the python artifact
building, because we'll push the whole folder for a release.