Cleanup artifact_version
Apt artifact building was using artifacting vars, which was not
rpc_release, and rpc_release was used for testing, which lead to
failures.  Using our (now standard) rpc_release everywhere,
we should be able to test and build artifacts.

To avoid duplication of work for the definition of rpc_release,
I kept the derive artifact version.

However, because of the derive artifact version shouldn't read
the git details because it's now set in a static file, we should
simplify its content.

That's what's done here.