Fix python artifact build for ceilometer/octavia
The python artifact build for octavia fails as there is no
git source information provided to match the presence of
the ansible role inclusion in ansible-role-requirements.

The python artifact build for ceilometer fails as the wheel
build process does not properly understand how to process
the optional dependency requirements.

The Mitaka upper constraint for elasticsearch is removed
so that the constraint from Newton is applied instead.