Add python artifact build tooling
In order to ensure that the python build consumes the
apt artifacts that have been prepared, the configuration
is implemented to consume the apt artifacts by default.

Using an almost standard AIO build process, an AIO is
prepared but only the repo container is built. The repo
container is used to build the python artifacts.

Once the build is complete, the artifacts are uploaded.